+++ CHANGES.md
@@ -11,6 +11,7 @@
         - weapon skills, eg. `pistol 4`, `mesmer gs 3`
         - profession skills, eg. `f3`, `scrapper f5`
         - engineer toolbelt skills, eg. `elixir h toolbelt`, `bk tb`
+    - `entity.Skill`: add `is_flipover`, `filter_is_main`
 - `textbody`: new module
 - `defnfile`:
     - parse `usage` and `notes` sections